ImplicitQueue | Add pushD_cost_worstcase
- Change `Exact_Queue` to take only one type argument. See the comment
for the rationale.
- Change the definition of `pushD`. Matching on two values at once leads
to case explosion in proofs due to the way Coq translates it.
on `q` at the outer level may seem dubious, but it makes sense. A demand
can be derived from `q` via `exact`, but there is no sensible way to go from
a `q` to `outD`. Matching on `q` at the outer level makes
proofs about `exact` easier.